+++ samples/trader-demo/src/main/kotlin/net/corda/traderdemo/TraderDemo.kt
@@ -2,7 +2,7 @@ package net.corda.traderdemo
 
 import joptsimple.OptionParser
 import net.corda.client.rpc.CordaRPCClient
-import net.corda.core.contracts.DOLLARS
+import net.corda.finance.DOLLARS
 import net.corda.core.utilities.NetworkHostAndPort
 import net.corda.core.utilities.loggerFor
 import net.corda.testing.DUMMY_BANK_A